Conflicts:
	application/admin/component/news/view/article/templates/default.html.php
	application/admin/public/theme/default/stylesheets/default.css
	component/news/database/row/article.php

+++ component/uploads/database/row/upload.php
@@ -63,12 +63,15 @@ class DatabaseRowUpload extends Library\DatabaseRowTable
     public function _importDistricts($data)
     {
         // Empty districts table
-        $this->getObject('com:districts.model.districts')->getRowset()->delete();
+        if($this->override)
+        {
+            $this->getObject('com:districts.model.districts')->getRowset()->delete();
+        }
 
         foreach($data as $item)
         {
             $row = $this->getObject('com:districts.database.row.district');
-            $row->id = $item['districts_district_id'];
+            $row->islp = $item['district_id'];
 
             if(!$row->load())
             {
@@ -111,7 +114,10 @@ class DatabaseRowUpload extends Library\DatabaseRowTable
     public function _importDistrictsofficers($data)
     {
         // Empty districts_districts_officers table
-        $this->getObject('com:districts.model.districts_officers')->getRowset()->delete();
+        if($this->override)
+        {
+            $this->getObject('com:districts.model.districts_officers')->getRowset()->delete();
+        }
 
         foreach($data as $item)
         {
@@ -129,22 +135,24 @@ class DatabaseRowUpload extends Library\DatabaseRowTable
     public function _importRelations($data)
     {
         // Empty districts_relations table
-        $this->getObject('com:districts.model.relations')->getRowset()->delete();
+        if($this->override)
+        {
+            $this->getObject('com:districts.model.relations')->getRowset()->delete();
+            $this->getObject('com:streets.model.relations')->table('districts_relations')->getRowset()->delete();
+        }
 
         foreach($data as $item)
         {
-            if(!array_key_exists('streets_street_id', $item))
+            if(!array_key_exists('streets_street_identifier', $item))
             {
-                $street = $this->getObject('com:streets.model.streets')->islp($item['islp'])->getRowset();
+                $street = $this->getObject('com:streets.model.streets')->islp($item['street_id'])->getRowset();
 
                 if(count($street))
                 {
-                    $item['streets_street_id'] = $street->top()->id;
+                    $item['streets_street_identifier'] = $street->top()->streets_street_identifier;
                 } else {
-                    $item['streets_street_id'] = '';
+                    continue;
                 }
-            } else {
-                $item['islp'] = '';
             }
 
             $parity = null;
@@ -152,28 +160,38 @@ class DatabaseRowUpload extends Library\DatabaseRowTable
                 case 'odd-even':
                 case 'Even+Oneven':
                 case 'Even/Oneven':
+                case 'Pair+Impair':
+                case 'Pair/Impair':
                     $parity = 'odd-even';
                     break;
                 case 'even':
                 case 'Even':
+                case 'Pair':
                     $parity = 'even';
                     break;
                 case 'odd':
                 case 'Oneven':
+                case 'Impair':
                     $parity = 'odd';
                     break;
             }
 
+            $item['districts_district_id'] = $this->getObject('com:districts.model.district')->islp($item['district_id'])->getRow()->id;
             $item['range_parity'] = $parity;
-            $item['id'] = sha1($item['districts_district_id'].$item['islp'].$item['streets_street_id'].$item['range_start'].$item['range_end'].$item['range_parity']);
 
             // Add row to districts_relations table when ID is unique
             $row = $this->getObject('com:districts.database.row.relation');
-            $row->id = $item['id'];
-            if(!$row->load())
-            {
-                $row->setData($item);
-                $row->save();
+            $row->setData($item);
+            $row->save();
+
+            // Add row to streets_relations
+            $relation = $this->getObject('com:streets.database.row.relation');
+            $relation->streets_street_identifier    = $item['streets_street_identifier'];
+            $relation->row                          = $row->id;
+            $relation->table                        = 'districts_relations';
+
+            if(!$relation->load()) {
+                $relation->save();
             }
         }
     }
@@ -188,17 +206,15 @@ class DatabaseRowUpload extends Library\DatabaseRowTable
             $city->load();
 
             //Get the street
-            $street = $this->getObject('com:streets.database.row.streets');
-            $street->title = $item['title'];
-            $street->streets_city_id = $city->streets_city_id;
+            $street = $this->getObject('com:streets.model.streets')->no_islp('1')->search($item['title'])->city($city->streets_city_id)->getRowset();
 
-            if($street->load()){
-                //Check if street does not have a islp value
-                if(!$street->islp) {
-                    //Set ISLP value and save
-                    $street->islp = $item['islp'];
-                    $street->save();
-                }
+            //Did we found a street with no ISLP number?
+            if($street)
+            {
+                $islp       = $this->getObject('com:streets.database.row.streets_islps');
+                $islp->id   = $street->streets_street_identifier;
+                $islp->islp = $item['islp'];
+                $islp->save();
             }
         }
     }
@@ -232,7 +248,7 @@ class DatabaseRowUpload extends Library\DatabaseRowTable
                 $row = $this->getObject('com:news.database.row.article');
                 $row->id = $item['id'];
 
-                // Only save the attachments when the row is new
+                // Only save when the row is new
                 if(!$row->load())
                 {
                     $row->title = $item['title'];
@@ -244,23 +260,12 @@ class DatabaseRowUpload extends Library\DatabaseRowTable
                     $row->modified_on = $item['modified'];
                     $row->modified_by = $item['modified_by'];
                     $row->published = $item['state'];
+                    $row->published_on = $item['publish_up'];
 
                     $this->_clean($row, 'news', true);
-                } else {
-                    $row->title = $item['title'];
-                    $row->slug = $item['alias'];
-                    $row->introtext = stripslashes($item['introtext']);
-                    $row->fulltext = stripslashes($item['fulltext']);
-                    $row->created_on = $item['created'];
-                    $row->created_by = '1';
-                    $row->modified_on = $item['modified'];
-                    $row->modified_by = $item['modified_by'];
-                    $row->published = $item['state'];
 
-                    $this->_clean($row, 'news', false);
+                    $row->save();
                 }
-                
-                $row->save();
             }
         }
     }
@@ -422,6 +427,12 @@ class DatabaseRowUpload extends Library\DatabaseRowTable
             if(empty($html)) {
                 continue;
             }
+
+            $html = preg_replace("/<wbr ?>/", "", $html);
+            $html = preg_replace("/<wbr ?\/>/", "", $html);
+
+            $pattern = "/<a href=\"mailto:[a-z0-9\"' =:&;\-_%@]+>(\b[A-Z0-9._%+-]+@[A-Z0-9.-]+\.[A-Z]{2,4}\b)<\/a>/i";
+            $html    = preg_replace($pattern, '<a href="mailto:$1">$1</a>', $html);
 
             $config = array(
                 'indent'         => true,
